Drop the Rails version of the ProGit book
There is a pre-rendered version of the ProGit book in this repository.
We will do the same in the upcoming static site that is rendered via
Hugo, but the contents will be slightly different.

So let's just remove this version and start from a clean slate before
introducing support code to pre-render the ProGit Book's pages as HTML
ready to be processed via Hugo.

Signed-off-by: Matt Burke <spraints@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>