Use relative URLs
As per https://ben.balter.com/jekyll-style-guide/config/#url, it is
highly recommended to use relative URLs where possible.

This is particularly necessary to make this here site work in GitHub
Pages because we want https://git.github.io/git-scm.com/ to render
correctly, even if the base URL is not at the top-level.

This trick was performed by:

	sed -i "s,\\(href=\"\|src=\"\)/\\([^\"]*\\),\\1{{ '\2' | relative_url }},g" $(git ls-files \*.html)

and manually fixing up e.g. the favicon in `_layouts/default.html`
(which is recorded with single quotes) and the collections URLs (which
already use the `{{ ... }}` syntax).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>