Use the base URL in partial .scss files
To support testing in forks of https://github.com/git-scm.com/, where
the site would not be located at the top-level directory (e.g.
https://user.github.io/git-scm.com/), let's be careful to include the
correct base URL in the CSS files.

This is a bit tricky, and needs `$baseurl` to be defined in
`application.scss`, and then this command was run:

	sed -i -e "s,\\(url(\|background-image-2x(\)\"/\\([^\"]*\\),\\1\$baseurl + \"\\2,g" \
		$(git ls-files _sass/\*.scss)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>