Convert 'About' pages to Hugo
This changes the pages so that they render correctly using Hugo.

An `_index.html` is added to redirect to the first 'About' page to keep
the current behavior; This cannot be done via `aliases:` in the front
matter of that page because we want a "list"-style page to redirect,
while `aliases:` only works if both source and target are "single"-style
pages.

Signed-off-by: Victoria Dye <vdye@github.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>