Migrate the 'Documentation' landing page to Hugo
This moves the information about the videos to the pages themselves, and
adjusts the landing page so that Hugo renders it as desired.

The video part of the logic is complicated enough to require
functionality that Hugo does not offer in `content/` pages; That part is
refactored into the `videos` shortcode, where the functionality _is_
available.

Co-authored-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Victoria Dye <vdye@github.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>