docs: prepare for `index.rake` to become a proper Ruby script
In the Jekyll world, we won't use Rake, but instead run things as regular
scripts. In preparation, rename the `index.rake` file accordingly.

Doing this before modifying the file will make it not only easier to follow the
development better, but also allows for rebasing to newer upstream revisions
more painlessly.

Signed-off-by: Victoria Dye <vdye@github.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>