update-book2: also include the images
With this commit, the script copies over the image contents and also
adjusts the links so that they work in forks of
https://github.com/git/git-scm.com/.

Be lenient when an image is referenced that does not exist. Do print
out an error in that instance, and preparing for updating all of
translations of the ProGit book in GitHub workflows, print it in the
GitHub workflow command syntax that will show the error on the summary
page. As of time of writing, such broken image links exist in the
following translations: cs, id, pl, tl and uz.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>