Use a pre-built search index
The `search-data.json` parsing takes longer than desirable on the client
side (and fetching a ~80MB file isn't exactly snappy either); Help this
by pre-generating it on the server side.

To help reduce the size even further, let's exclude the past versions of
the manual pages as well as the translated ones.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>