fixup! TODOs
Going to Hugo would cost us the convenience of using GitHub's pre-built
`pages-build-deployment` workflow, and we'd have to go through those
steps manually.

Happily, this is not as big of a deal as one might think: just maintain
`gh-pages` as a _separate_ commit history, containing the HTML site (no
Jekyll, no Hugo, all post-rendered), and update it via GitHub workflows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>