Define and use sections in a flexible way
Sections are defined in a variety of ways. Manual pages, for example,
are implicitly documentation because they are in the `content/docs/`
directory.

Let's define a Hugo variable `$section` to hold the name of the current
section. For good measure, also store this information in the scratch
space that can be accessed by Hugo's partials.

Use that variable in the `sidebar` to determine what section is
currently active (after Hugo'fying the links).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>