search: exclude older manual page versions
Old manual page versions are almost always uninteresting. So let's avoid
showing them in the search results, especially since they would likely
clutter the results page for pages that changed often in the past.

This feat is accomplished by skipping the `data-pagefind-body` attribute
when looking at a manual page version that is not the latest.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>