Move logos page to the new location
Instead of letting the Rails App render the logos page on the fly, we
are about to let Hugo render it into a static site. To that end, let's
move the `.html.erb` file into the location where Hugo expects it, and
then adapt the content subsequently. Splitting this into several commits
will help Git recognize this as a move; Doing it in a single commit
would be mistakend for a delete/add instead and make the transition
harder to follow.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>