search: reintroduce the categories in the live search
In the existing Rails app, when the user types a search term into the
search box, the results are sorted into the categories "Reference" and
"Book". Let's do the same with the Pagefind-based search.

We do this by specifically adding metadata indicating the category of
manual pages as well as for the sections of the book (for details, see
https://pagefind.app/docs/metadata/). Then, we use that information to
sort the results into the correct table row. Due to the `async` nature
of the search results coming in, we need to be
careful to process them in order when doing that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>