docs: handle `linkgit:curl[1]` gracefully
It is a mistake to use a `linkgit` for a non-Git command. There were
such mistakes in older Git versions, fixed in 1cca17dfff48
(Documentation: fix linkgit references, 2016-05-04).

Since we want to build older Git versions' manual pages, and since it
was just one, consistent mistake where a page wanted to link to the
`curl` documentation, be nice and actually link there.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>