playwright: add a few platform-specific tests
The Git home page displays some platform-dependent behavior. With this
commit, we test that by pretending to run on the three major operating
systems: Windows, macOS and Linux. Specifically, we verify that:

- the correct download is suggested, based on the current browser

- a subset of the GUIs are shown based on the platform indicated by the
  current browser

Pretending to be running on a given platform is a bit tricky because the
`navigator.userAgent`/`navigator.platform` attributes are used by the
Git home page (via `session.min.js`) to determine what operating system
the user is running.

These can be easily overridden on most browser, using the
`Object.defineProperty()` method.

However, in my tests, Chromium on Linux (and only on that operating
system) simply ignored those calls and still reported that it is running
on Linux.

To work around this, we need to use the Chrome DevTools Protocol, via
the `BrowserContext.newCDPSession()` method; For more details see
https://playwright.dev/docs/api/class-browsercontext#browser-context-new-cdp-session

Obviously, this can _only_ be used with Chromium/Chrome and won't work
on, say, Firefox or Safari. That is the reason for the somewhat ugly
`pretendPlatform()` function we need to use here.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>