playwright: verify that URLs with question marks work
Some book sections' titles end in a question mark. In URLs, question
marks are interpreted as the separator between the path part and the
query string. This matters because the book sections' titles are used to
construct their URLs.

We want to follow Postel's Law and accept even URLs with non-URI-encoded
question marks, gracefully redirecting to the correct URL. Add code to
the Playwright-based tests to that extent.

While at it, also verify that the redirection uses proper URI-encoding
for non-ASCII characters; This requires a minor improvement of the
intentionally very simplistic `serve-public.js` script.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>